steam-bridge: compile the generated code with warnings-as-errors
The Steam per-tag object libraries compile only our generated thunks/proxies
(plus the runtime headers), yet they opted out of the project's warnings-as-errors
via disable_warnings -- so real issues in the generated code went unnoticed and the
GCC build spewed noise. Hold the generated code to the same standard as the rest of
src instead:

- Mark the Valve SDK headers as SYSTEM includes on the per-tag targets, keeping the
  generated headers on the normal path. This suppresses the SDK's own warnings and
  the cross-snapshot -Wodr noise under LTO, while our generated code still warns.
- Drop the disable_warnings opt-out on the per-tag libs so they inherit -Werror.
  On MSVC, lower external-header warnings (/external:W0), define
  _CRT_SECURE_NO_WARNINGS, and suppress C4743/C4744 -- the MSVC analogue of -Wodr,
  where identically-named ISteam* types differ across snapshots and MSVC flags the
  vtable/data size mismatch at LTCG link time. The hand-written base TUs are unchanged.

Fix the warning sources this surfaced:
- guest proxy stub bodies for non-marshallable methods named their parameters
  without using them; cast them to void.
- null_ptr params (e.g. double-pointer filter lists the host substitutes with
  nullptr) were likewise named but unused; cast to void.
- the invoker's out_pos_ (uint32_t) took `+= n` with a size_t n, narrowing on
  64-bit (MSVC C4267); cast explicitly.
- a real -Wtautological-compare in steam_host_runtime.hpp: the reader's overflow
  guard `p + n >= p` is always true (pointer overflow is UB); use the equivalent
  non-overflowing `n <= end - p`.

Also make sogen_target_disable_warnings actually silence warnings on GCC/Clang
(-w), which it only did for MSVC before.

+++ src/samples/steam-shim/steam_shim_runtime.hpp
@@ -125,7 +125,7 @@ namespace sogen::steam_shim
             {
                 std::memcpy(dst, out_.data() + out_pos_, n);
             }
-            out_pos_ += n;
+            out_pos_ += static_cast<uint32_t>(n);
         }
 
         void call()
